requirements.txt
langchain>=1.2.11
langchain-community>=0.4.1
langchain-ollama>=1.0.1
langchain-openai>=1.1.11
faiss-cpu>=1.13.2
python-dotenv>=1.2.2